requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Mail==0.9.1
Flask-SocketIO==5.1.1
Flask-DebugToolbar==0.13.1
python-dotenv==0.19.2
pymongo[srv]==3.12.1
requests==2.26.0
